chore: Switch to commonjs module resolution
Hi, i suggest to switch to commonjs as module resolution. Currently it is necessary to reconfigure jest, webpack and other build processes to transpile the contrast library aswell. This should not be necessary, a npm package should be shipped in a way that it is not necessary to transpile it again, this should be do before the publishing.
I hope my change is fine for you.